refactor(admin): MobileBannerSection 컴포넌트 분리 및 파일 크기 검증 추가
